Refactor: Extracted AST evaluation to separate subcrate
