+++ tests/Cargo.toml
@@ -10,6 +10,7 @@ license.workspace = true
 werk-runner.workspace = true
 werk-parser.workspace = true
 werk-fs.workspace = true
+werk-eval.workspace = true
 anyhow = "1"
 parking_lot.workspace = true
 ahash.workspace = true
