+++ tests/test_eval.rs
@@ -1,6 +1,7 @@
 use tests::mock_io::*;
+use werk_eval::{EvalError, ShellCommandLine};
 use werk_parser::parser::{Input, parse};
-use werk_runner::{Error, EvalError, ShellCommandLine, eval};
+use werk_runner::Error;
 use werk_util::DiagnosticFileId;
 use winnow::Parser as _;
 
@@ -14,7 +15,7 @@ fn command_argument_splitting() {
 
     // Simple literal command.
     let expr = parse.parse(Input::new(r#""a""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
@@ -25,7 +26,7 @@ fn command_argument_splitting() {
 
     // Simple interpolation.
     let expr = parse.parse(Input::new(r#""{foo}""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
@@ -36,7 +37,7 @@ fn command_argument_splitting() {
 
     // One literal argument
     let expr = parse.parse(Input::new(r#""a b""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
@@ -47,7 +48,7 @@ fn command_argument_splitting() {
 
     // Expand list without expansion as the first entry.
     let expr = parse.parse(Input::new(r#""a {abc}""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
@@ -58,7 +59,7 @@ fn command_argument_splitting() {
 
     // Expand list with expansion as separate arguments.
     let expr = parse.parse(Input::new(r#""a {abc*}""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
@@ -69,7 +70,7 @@ fn command_argument_splitting() {
 
     // ... unless there is a join separator
     let expr = parse.parse(Input::new(r#""a {abc,*}""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
@@ -80,7 +81,7 @@ fn command_argument_splitting() {
 
     // ... or the argument is quoted.
     let expr = parse.parse(Input::new(r#""a \"{abc*}\"""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
@@ -91,7 +92,7 @@ fn command_argument_splitting() {
 
     // Support single-quoting, for things like `sh -c 'foo bar'`.
     let expr = parse.parse(Input::new(r#""a -c 'a b'""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
@@ -102,7 +103,7 @@ fn command_argument_splitting() {
 
     // Argument expansion within single quotes
     let expr = parse.parse(Input::new(r#""a -c '{abc*}'""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
@@ -113,7 +114,7 @@ fn command_argument_splitting() {
 
     // Quotes in interpolated variables do not terminate a quoted argument.
     let expr = parse.parse(Input::new(r#""a \"{q}\"""#)).unwrap();
-    let cmd = eval::eval_shell_command(workspace, &expr, file).unwrap();
+    let cmd = werk_eval::eval_shell_command(workspace, &expr, file).unwrap();
     assert_eq!(
         cmd.value,
         ShellCommandLine {
