+++ tests/test_expressions.rs
@@ -1,4 +1,4 @@
-use werk_runner::Value;
+use werk_eval::Value;
 
 use stringleton::Symbol;
 use tests::mock_io::*;
