+++ werk-cli/Cargo.toml
@@ -33,6 +33,7 @@ werk-fs.workspace = true
 werk-parser.workspace = true
 werk-runner.workspace = true
 werk-util.workspace = true
+werk-eval.workspace = true
 anstyle = "1.0.10"
 serde.workspace = true
 serde_json = "1.0.137"
