+++ werk-eval/evaluated.rs
@@ -1,0 +1,110 @@
+use crate::{Used, UsedVariable, Value};
+
+/// Evaluated value, which keeps track of "outdatedness" with respect to cached
+/// build variables. Build recipes may use the outdatedness information to
+/// determine if some expression changed in a way that should cause a rebuild to
+/// occur, like the result of a glob pattern, or the result of a `which`
+/// expression (which is outdated if executable program path changed between
+/// runs).
+#[derive(Clone, Debug)]
+pub struct Eval<T = Value> {
+    pub value: T,
+    /// When the evaluated value can impact outdatedness (like a glob
+    /// expression), this is the variables used during evaluation.
+    pub used: Used,
+}
+
+impl<T> Eval<T> {
+    /// An inherent value that does not depend on any external variables.
+    pub const fn inherent(value: T) -> Self {
+        Self {
+            value,
+            used: Used::none(),
+        }
+    }
+
+    pub fn using_var(value: T, used: UsedVariable) -> Self {
+        Self::using_vars(value, [used])
+    }
+
+    pub fn using_vars(value: T, used: impl IntoIterator<Item = UsedVariable>) -> Self {
+        Self {
+            value,
+            used: Used::from_iter(used),
+        }
+    }
+
+    pub fn map<U>(self, f: impl FnOnce(T) -> U) -> Eval<U> {
+        Eval {
+            value: f(self.value),
+            used: self.used,
+        }
+    }
+
+    pub fn as_deref(&self) -> &T::Target
+    where
+        T: std::ops::Deref,
+    {
+        &self.value
+    }
+}
+
+impl<T> AsRef<T> for Eval<T> {
+    fn as_ref(&self) -> &T {
+        &self.value
+    }
+}
+
+impl<T: Default> Default for Eval<T> {
+    fn default() -> Self {
+        Self {
+            value: Default::default(),
+            used: Used::none(),
+        }
+    }
+}
+
+impl<T> Eval<&T> {
+    #[must_use]
+    pub fn cloned(&self) -> Eval<T>
+    where
+        T: Clone,
+    {
+        Eval {
+            value: self.value.clone(),
+            used: self.used.clone(),
+        }
+    }
+}
+
+impl<T> Eval<Option<T>> {
+    pub fn transpose(self) -> Option<Eval<T>> {
+        self.value.map(|value| Eval {
+            value,
+            used: self.used,
+        })
+    }
+}
+
+impl<T, E> Eval<Result<T, E>> {
+    pub fn transpose_result(self) -> Result<Eval<T>, E> {
+        self.value.map(|value| Eval {
+            value,
+            used: self.used,
+        })
+    }
+}
+
+impl<T> std::ops::Deref for Eval<T> {
+    type Target = T;
+
+    fn deref(&self) -> &Self::Target {
+        &self.value
+    }
+}
+
+impl<T> std::ops::DerefMut for Eval<T> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.value
+    }
+}
