+++ werk-eval/lib.rs
@@ -1,0 +1,39 @@
+mod command_line;
+mod constants;
+mod env;
+mod error;
+mod eval;
+mod eval_string;
+mod evaluated;
+mod glob_settings;
+mod io;
+mod messenger;
+mod pattern;
+mod recipe;
+mod run_command;
+mod scope;
+mod task_id;
+mod used;
+mod value;
+mod warning;
+
+pub use command_line::*;
+pub use constants::*;
+pub use env::*;
+pub use error::*;
+pub use eval::*;
+pub use eval_string::*;
+pub use evaluated::*;
+pub use glob_settings::*;
+pub use io::*;
+pub use messenger::*;
+pub use pattern::*;
+pub use recipe::*;
+pub use run_command::*;
+pub use scope::*;
+pub use task_id::*;
+pub use used::*;
+pub use value::*;
+pub use warning::*;
+
+stringleton::enable!();
