+++ werk-planner/lib.rs
@@ -1,14 +1,9 @@
-pub fn add(left: u64, right: u64) -> u64 {
-    left + right
-}
+mod recipe;
+mod task_graph;
+mod task_id;
+mod task_name;
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_works() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
-    }
-}
+pub use recipe::*;
+pub use task_graph::*;
+pub use task_id::*;
+pub use task_name::*;
