+++ werk-runner/outdatedness.rs
@@ -4,14 +4,10 @@ use std::{
 };
 
 use stringleton::Symbol;
+use werk_eval::{TaskId, Used, UsedVariable};
 use werk_fs::{Absolute, SymPath};
 
-use crate::{
-    TaskId, Workspace,
-    cache::TargetOutdatednessCache,
-    eval::{Used, UsedVariable},
-    ir,
-};
+use crate::{Workspace, cache::TargetOutdatednessCache};
 
 /// A reason why a variable or recipe is "outdated".
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -170,7 +166,7 @@ impl<'a> OutdatednessTracker<'a> {
     pub fn new(
         workspace: &'a Workspace,
         cache: Option<&'a TargetOutdatednessCache>,
-        recipe: &ir::BuildRecipe,
+        recipe: &werk_eval::BuildRecipe,
         target_mtime: Option<std::time::SystemTime>,
     ) -> Self {
         let mut outdatedness = Outdatedness::unchanged();
