+++ werk-util/io_error.rs
@@ -1,0 +1,43 @@
+use std::sync::Arc;
+
+/// Clonable `std::io::Error`
+#[derive(Clone)]
+pub struct IoError {
+    pub error: Arc<std::io::Error>,
+}
+
+impl From<std::io::Error> for IoError {
+    #[inline]
+    fn from(error: std::io::Error) -> Self {
+        Self {
+            error: Arc::new(error),
+        }
+    }
+}
+
+impl std::fmt::Debug for IoError {
+    #[inline]
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::fmt::Debug::fmt(&*self.error, f)
+    }
+}
+
+impl std::fmt::Display for IoError {
+    #[inline]
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::fmt::Display::fmt(&*self.error, f)
+    }
+}
+
+impl PartialEq for IoError {
+    #[inline]
+    fn eq(&self, other: &Self) -> bool {
+        Arc::ptr_eq(&self.error, &other.error) || self.error.kind() == other.error.kind()
+    }
+}
+
+impl std::error::Error for IoError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        self.error.source()
+    }
+}
