+++ werk-util/lib.rs
@@ -1,10 +1,13 @@
 pub mod cancel;
 mod diagnostic;
+pub mod hash128;
+mod io_error;
 mod os_str;
 mod semantic_hash;
 mod span;
 
 pub use diagnostic::*;
+pub use io_error::*;
 pub use os_str::*;
 pub use semantic_hash::*;
 pub use span::*;
